+++ packages/edit-widgets/src/components/more-menu/tools-more-menu-group.js
@@ -1,0 +1,21 @@
+/**
+ * External dependencies
+ */
+import { isEmpty } from 'lodash';
+
+/**
+ * WordPress dependencies
+ */
+import { createSlotFill } from '@wordpress/components';
+
+const { Fill: ToolsMoreMenuGroup, Slot } = createSlotFill(
+	'EditWidgetsToolsMoreMenuGroup'
+);
+
+ToolsMoreMenuGroup.Slot = ( { fillProps } ) => (
+	<Slot fillProps={ fillProps }>
+		{ ( fills ) => ! isEmpty( fills ) && fills }
+	</Slot>
+);
+
+export default ToolsMoreMenuGroup;